+++ src/manager/cron/frequent.js
@@ -1,0 +1,3 @@
+const run = require('./runner.js');
+
+module.exports = async (options) => run('frequent', options);